pyproject.toml
[project]
name = "backend"
version = "2.0.0"
description = "AVP Conecta - API REST para Gestão de Eventos Acadêmicos"
requires-python = ">=3.12"
dependencies = [
    # Web Framework & Server
    "fastapi>=0.128.8",
    "uvicorn[standard]>=0.39.0",
    # Database
    "SQLAlchemy>=2.0.48",
    "PyMySQL>=1.1.2",
    "alembic>=1.16.5",
    # Security & Auth
    "python-jose[cryptography]>=3.5.0",
    "passlib[argon2]>=1.7.4",
    # Validation & Settings
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "email-validator>=2.3.0",
    "python-dotenv>=1.2.1",
    # Utils
    "fpdf2>=2.8.4",
    "python-multipart>=0.0.20",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "defusedxml>=0.7.1",
    "PyYAML>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
